fix integer type warnings

+++ extras/rapidfuzz_amalgamated.hpp
@@ -1,7 +1,7 @@
 //  Licensed under the MIT License <http://opensource.org/licenses/MIT>.
 //  SPDX-License-Identifier: MIT
 //  RapidFuzz v1.0.2
-//  Generated: 2023-10-12 23:14:55.535235
+//  Generated: 2023-10-21 19:41:24.603454
 //  ----------------------------------------------------------
 //  This file is an amalgamation of multiple different files.
 //  You probably shouldn't edit it directly.
@@ -5747,10 +5747,11 @@ auto jaro_similarity_prepare_bound(const VecType* s1_lengths, Range<InputIt>& s2
         // side-note: we know only the first 8 bit are actually used
         alignas(alignment) std::array<VecType, vec_width> lens_;
         s1_lengths_simd.store(lens_.data());
-        for (int i = 0; i < vec_width; ++i)
+        for (size_t i = 0; i < vec_width; ++i)
             if (lens_[i] > maxLen) maxLen = lens_[i];
 
-        bounds.maxBound = (s2.size() > maxLen) ? s2.size() : maxLen;
+        bounds.maxBound =
+            (s2.size() > static_cast<ptrdiff_t>(maxLen)) ? s2.size() : static_cast<ptrdiff_t>(maxLen);
         bounds.maxBound /= 2;
         if (bounds.maxBound > 0) bounds.maxBound--;
